// src/app/models/custom-text.model.ts
export class CustomText {
  TextId: number;
  PageId: number;
  color: string;
  fontSize: string;
  textAlign: string;
  paddingTop: string;
  paddingBottom: string;
  paddingLeft: string;
  paddingRight: string;
  marginTop: string;
  marginBottom: string;
  marginLeft: string;
  marginRight: string;
  lineHeight: string;
  fontFamily: string;
  border: string;
  borderStyle: string;
  isText: string;
  TextBody: string;
  borderBottom: string;
  borderBottomColor: string;
  borderBottomStyle: string;
  borderBottomWidth: string;
  borderColor: string;
  borderLeftStyle: string;
  borderLeftColor: string;
  borderLeftWidth: string;
  borderRight: string;
  borderRightColor: string;
  borderRightStyle: string;
  borderTop: string;
  borderTopColor: string;
  borderTopStyle: string;
  borderTopWidth: string;
  borderWidth: string;
  clear: string;
  clip: string;
  cursor: string;
  font: string;
  fontVariant: string;
  fontWeight: string;
  height: string;
  left: string;
  letterSpacing: string;
  listStyle: string;
  listStyleImage: string;
  listStyleType: string;
  margin: string;
  pageBreakAfter: string;
  pageBreakBefore: string;
  position: string;
  strokeDasharray: string;
  strokeDashoffset: string;
  strokeWidth: string;
  textDecoration: string;
  textIndent: string;
  textTransform: string;
  top: string;
  right: string;
  bottom: string;
  verticalAlign: string;
  visibility: string;
  width: string;
  zIndex: string;
  display: string;
  cssFloat: string;
  filter: string;
  backgroundColor: string;
  backgroundPosition: string;
  backgroundRepeat: string;
  lineBreak: string;
  opacity: string;
  outline: string;
  outlineColor: string;
  outlineOffset: string;
  outlineStyle: string;
  outlineWidth: string;
  padding: string;
  paddingBlock: string;
  paddingBlockEnd: string;
  paddingBlockStart: string;
  paddingInlineEnd: string;
  paddingInlineStart: string;
  pageBreakInside: string;
  rotate: string;
  textAlignLast: string;
  textCombineUpright: string;
  textDecorationColor: string;
  textDecorationLine: string;
  textDecorationSkipInk: string;
  textDecorationStyle: string;
  textDecorationThickness: string;
  textEmphasis: string;
  textEmphasisColor: string;
  textEmphasisPosition: string;
  textEmphasisStyle: string;
  textJustify: string;
  textOrientation: string;
  textOverflow: string;
  textRendering: string;
  textShadow: string;
  textUnderlineOffset: string;
  textUnderlinePosition: string;
  transformBox: string;
  transformOrigin: string;
  transition: string;
  transitionDelay: string;
  transitionDuration: string;
  transitionProperty: string;
  transitionTimingFunction: string;
  translate: string;
  whiteSpace: string;
  wordBreak: string;
  wordSpacing: string;
  wordWrap: string;
  writingMode: string;
  flex: string;
  alignContent: string;
  alignItems: string;
  alignSelf: string;
  animation: string;
  customField: string;
}